+++ models/archived_jobs/archived_jobs_test.go
@@ -16,6 +16,6 @@ func ExampleCreate() {
 
 func ExampleGet() {
 	id, _ := types.NewPrefixUUID("job_6740b44e-13b9-475d-af06-979627e0e0d6")
-	aj, _ := Get(id)
+	aj, _ := Get(context.TODO(), id)
 	fmt.Println(aj.ID.String())
 }